+++ src/manifest.m
@@ -18,6 +18,7 @@
 #include <sys/wait.h>
 #include <sys/stat.h>
 #include <sys/types.h>
+#include <sys/sysctl.h>
 #include <sys/socket.h>
 #include <sys/un.h>
 #include <netinet/in.h>