+++ src/yabai.c
@@ -23,7 +23,7 @@
 
 #define MAJOR  5
 #define MINOR  0
-#define PATCH  3
+#define PATCH  4
 
 struct event_loop g_event_loop;
 void *g_workspace_context;